# Demo/eFishery Demo.postman_collection.json-Swagger20.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "eFishery Demo",
    "contact": {}
  },
  "host": "localhost:9000",
  "basePath": "/api/e-commerce/v1",
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/admin/roles": {
      "post": {
        "summary": "Post Role (user: admin; pass: efishery)",
        "tags": [
          "Role & User"
        ],
        "operationId": "PostRole(user:admin;pass:efishery)",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/PostRole%28user%3Aadmin%3Bpass%3Aefishery%29Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "basic": []
          }
        ]
      },
      "get": {
        "summary": "Get All Roles",
        "tags": [
          "Role & User"
        ],
        "operationId": "GetAllRoles",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "basic": []
          }
        ]
      }
    },
    "/users": {
      "post": {
        "summary": "Create User",
        "tags": [
          "Role & User"
        ],
        "operationId": "CreateUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "summary": "Get All Users & By Role Too",
        "tags": [
          "Role & User"
        ],
        "operationId": "GetAllUsers&ByRoleToo",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateUserRequest": {
      "title": "CreateUserRequest",
      "example": {
        "Name": "eFishery customer 3",
        "Email": "email customer 3",
        "Phone": "customer phone 3",
        "Address": "customer address 3",
        "NIK": "customer NIK 3",
        "Role_Ref": 2
      },
      "type": "object",
      "properties": {
        "Name": {
          "type": "string"
        },
        "Email": {
          "type": "string"
        },
        "Phone": {
          "type": "string"
        },
        "Address": {
          "type": "string"
        },
        "NIK": {
          "type": "string"
        },
        "Role_Ref": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "Name",
        "Email",
        "Phone",
        "Address",
        "NIK",
        "Role_Ref"
      ]
    },
...
  },
  "security": [],
...
}